// viz/charts/network-tools/types.ts
export type LabelPosOptions = typeof labelPosOptionNames[number]

export type GraphNode = {
  _id: string | number
  _label?: string
  _labelColor?: string
  _labelFont?: string
  _labelFontWeight?: string
  _labelYOffset?: number
  _backgroundColor?: string
  _backgroundSize?: number
  _backgroundShape?: 'circle' | 'hexagon'
  _color?: string
  _shape?: string
  _nodeType?: string
  _show?: boolean
  _showLabel?: boolean
  _labelPos?: LabelPosOptions
  _fontSize?: number
  _icon?: string
  _size?: number
  x?: number
  y?: number
  fx?: number
  fy?: number
}

export type GraphLink = {
  source: GraphNode['_id'] | GraphNode
  target: GraphNode['_id'] | GraphNode
  value?: number
  color?: string
  _linkType?: string
}

export type AppGraphData = {
  nodes: GraphNode[]
  links: GraphLink[]
}